+++ requirements.txt
@@ -5,7 +5,6 @@ pycodestyle==2.11.1
 pydocstyle==6.3.0
 pylint==3.1.0
 tox==4.15.0
-vulture==2.11
 types-PyYAML==6.0.12.20240311
 databricks-connect==14.3.1
 sqlglot==23.13.7
