+++ tox.ini
@@ -15,7 +15,6 @@ commands =
     black . -l 120 --check
     pydocstyle .
     pylint --load-plugins=databricks.labs.pylint.all notebooks sqlmesh_projects
-    vulture notebooks
 
 [pycodestyle]
 max-line-length = 120
